Proposed fix for #72
Fixes issue #72 where quest objectives outside of the quest zone are not shown.  
Fixed by checking for zone specific coordinates first then trying other zones if the coordinates are out of bounds.  
Then filters out results that are in other zones only if there are results that are in the quest zone.